refactor(lotto): lottoNumbers 를 도메인으로 변경, validation 위치 이동
