--- src/main/kotlin/lotto/dto/LottoNumbers.kt
@@ -1,4 +1,0 @@
-package lotto.dto
-
-@JvmInline
-value class LottoNumbers(val lottoNumbers: List<Int>)
